+++ lib/ruby-progress/cli/ripple_cli.rb
@@ -139,9 +139,6 @@ module RippleCLI
       Signal.trap('TERM') { stop_requested = true }
       Signal.trap('HUP')  { stop_requested = true }
 
-      job_dir = RubyProgress::Daemon.job_dir_for_pid(pid_file)
-      job_thread = Thread.new { process_daemon_jobs_for_rippler(job_dir, rippler, options) }
-
       rippler.advance until stop_requested
     ensure
       RubyProgress::Utils.clear_line
@@ -184,49 +181,7 @@ module RippleCLI
       end
 
       # stop job thread and cleanup
-      job_thread&.kill
       FileUtils.rm_f(pid_file)
-    end
-  end
-
-  def self.process_daemon_jobs_for_rippler(job_dir, rippler, options)
-    RubyProgress::Daemon.process_jobs(job_dir) do |job|
-      jid = job['id'] || SecureRandom.uuid
-      log_path = begin
-        File.join(File.dirname(job_dir), "#{jid}.log")
-      rescue StandardError
-        nil
-      end
-
-      oc = RubyProgress::OutputCapture.new(
-        command: job['command'],
-        lines: options[:output_lines] || 3,
-        position: options[:output_position] || :above,
-        log_path: log_path
-      )
-      oc.start
-
-      rippler.instance_variable_set(:@output_capture, oc)
-      oc.wait
-      captured = oc.lines.join("\n")
-      exit_status = oc.exit_status
-      rippler.instance_variable_set(:@output_capture, nil)
-
-      success = exit_status.to_i.zero?
-      if job['message']
-        RubyProgress::Utils.display_completion(
-          job['message'],
-          success: success,
-          show_checkmark: job['checkmark'] || false,
-          output_stream: :stdout,
-          icons: { success: options[:success_icon], error: options[:error_icon] }
-        )
-      end
-
-      { 'exit_status' => exit_status, 'output' => captured, 'log_path' => log_path }
-    rescue StandardError
-      # ignore per-job errors; process_jobs will write result
-      nil
     end
   end